+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Allow, but raise warning when specifying a CRS for the geometry passed to `aggregate_spatial` and `mask_polygon`,
+  which is non-standard/experimental feature, only supported by specific back-ends 
+  ([#204](https://github.com/Open-EO/openeo-python-client/issues/204))
+
 ### Changed
 
 ### Removed